Update README.md
Hi!! 

This is a great list, but you are missing cowswap.exchange which is the first dapp to be developed on top of Gnosis Protocol v2. It was also the first live DEX that was able to offer MEV protected trades without the need for Flashbots. Hope you can include it